@@ -8,7 +8,7 @@
 
 #define MOBILE_PLATFORM_IOS 1
 
-#elif defined(WINAPI_FAMILY) && ((WINAPI_FAMILY == WINAPI_FAMILY_PC_APP) || (WINAPI_FAMILY == WINAPI_FAMILY_PHONE_APP))
+#elif defined(WINAPI_FAMILY) && (WINAPI_FAMILY == WINAPI_FAMILY_PC_APP)
 
 #define MOBILE_PLATFORM_WINRT 1
 
