@@ -243,10 +243,11 @@ public:
      *
      *  @param  callback    the callback to execute
      */
-    Deferred &onSuccess(const SuccessCallback &callback)
+    inline Deferred &onSuccess(const SuccessCallback& callback) { return onSuccess(SuccessCallback(callback)); }
+    Deferred &onSuccess(SuccessCallback&& callback)
     {
         // store callback
-        _successCallback = callback;
+        _successCallback = std::move(callback);
 
         // allow chaining
         return *this;
@@ -262,13 +263,14 @@ public:
      *
      *  @param  callback    the callback to execute
      */
-    Deferred &onError(const ErrorCallback &callback)
+    inline Deferred &onError(const ErrorCallback& callback) { return onError(ErrorCallback(callback)); }
+    Deferred &onError(ErrorCallback&& callback)
     {
         // store callback
-        _errorCallback = callback;
+        _errorCallback = std::move(callback);
 
         // if the object is already in a failed state, we call the callback right away
-        if (_failed) callback("Frame could not be sent");
+        if (_failed) _errorCallback("Frame could not be sent");
 
         // allow chaining
         return *this;
@@ -289,13 +291,14 @@ public:
      *
      *  @param  callback    the callback to execute
      */
-    Deferred &onFinalize(const FinalizeCallback &callback)
+    inline Deferred &onFinalize(const FinalizeCallback& callback) { return onFinalize(FinalizeCallback(callback)); }
+    Deferred &onFinalize(FinalizeCallback&& callback)
     {
         // if the object is already in a failed state, we call the callback right away
         if (_failed) callback();
 
         // otherwise we store callback until it's time for the call
-        else _finalizeCallback = callback;
+        else _finalizeCallback = std::move(callback);
 
         // allow chaining
         return *this;
