@@ -76,10 +76,11 @@ public:
      *
      *  @param  callback    the callback to execute
      */
-    DeferredDelete &onSuccess(const DeleteCallback &callback)
+    inline DeferredDelete &onSuccess(const DeleteCallback& callback) { return onSuccess(DeleteCallback(callback)); }
+    DeferredDelete &onSuccess(DeleteCallback&& callback)
     {
         // store callback
-        _deleteCallback = callback;
+        _deleteCallback = std::move(callback);
         
         // allow chaining
         return *this;
@@ -89,10 +90,11 @@ public:
      *  Register the function that is called when the queue is deleted or purged
      *  @param  callback
      */
-    DeferredDelete &onSuccess(const SuccessCallback &callback)
+    inline DeferredDelete &onSuccess(const SuccessCallback& callback) { return onSuccess(SuccessCallback(callback)); }
+    DeferredDelete &onSuccess(SuccessCallback&& callback)
     {
         // call base
-        Deferred::onSuccess(callback);
+        Deferred::onSuccess(std::move(callback));
         
         // allow chaining
         return *this;
