@@ -76,10 +76,11 @@ public:
      *
      *  @param  callback    the callback to execute
      */
-    DeferredQueue &onSuccess(const QueueCallback &callback)
+    inline DeferredQueue &onSuccess(const QueueCallback& callback) { return onSuccess(QueueCallback(callback)); }
+    DeferredQueue &onSuccess(QueueCallback&& callback)
     {
         // store callback
-        _queueCallback = callback;
+        _queueCallback = std::move(callback);
         
         // allow chaining
         return *this;
@@ -89,10 +90,11 @@ public:
      *  Register the function that is called when the queue is declared
      *  @param  callback
      */
-    DeferredQueue &onSuccess(const SuccessCallback &callback)
+    inline DeferredQueue &onSuccess(const SuccessCallback& callback) { return onSuccess(SuccessCallback(callback)); }
+    DeferredQueue &onSuccess(SuccessCallback&& callback)
     {
         // call base
-        Deferred::onSuccess(callback);
+        Deferred::onSuccess(std::move(callback));
         
         // allow chaining
         return *this;
