@@ -47,9 +47,9 @@ protected:
 
     /**
      *  Number of bytes already processed
-     *  @var    uint32_t
+     *  @var    size_t
      */
-    uint32_t _skip = 0;
+    size_t _skip = 0;
 
 public:
     /**
@@ -128,7 +128,7 @@ public:
      *  @param  size
      *  @return char*
      */
-    const char *nextData(uint32_t size);
+    const char *nextData(size_t size);
 
     /**
      *  The checker may access private data
